+++ projects/angular-sdk/src/apis/hopper-cloud-airline/v1/model/cfarContractCustomer.ts
@@ -10,5 +10,6 @@
  * Do not edit the class manually.
  */
 
-export interface Unauthorized { 
+export interface CfarContractCustomer { 
+    id: string;
 }