+++ projects/angular-sdk/src/apis/hopper-cloud-airline/v1/model/createCfarContractCustomerRequest.ts
@@ -9,7 +9,13 @@
  * https://github.com/swagger-api/swagger-codegen.git
  * Do not edit the class manually.
  */
+import { CfarItinerary } from './cfarItinerary';
 
-export interface OfflineReconciliation { 
-    type: string;
+/**
+ * A create CFAR contract customer request
+ */
+export interface CreateCfarContractCustomerRequest { 
+    offerIds: Array<string>;
+    itinerary: CfarItinerary;
+    pnrReference?: string;
 }