+++ projects/angular-sdk/src/components/cfar-offer-banner/cfar-offer-banner.component.ts
@@ -1,6 +1,6 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { take } from 'rxjs/operators';
-import { CfarContract, CfarItinerary, CfarOffer, CreateCfarOfferRequest, RequestType } from '../../apis/hopper-cloud-airline/v1';
+import { CfarContractCustomer, CfarItinerary, CfarOfferCustomer, CreateCfarOfferCustomerRequest, RequestType } from '../../apis/hopper-cloud-airline/v1';
 import { GlobalComponent } from '../global.component';
 import { TranslateService } from '@ngx-translate/core';
 import { DateAdapter } from "@angular/material/core";
@@ -17,10 +17,10 @@ import { DialogUtils } from '../../utils/dialog.utils';
 })
 export class CfarOfferBannerComponent extends GlobalComponent implements OnInit {
 
-  public cheapestOffer!: CfarOffer;
+  public cheapestOffer!: CfarOfferCustomer;
   public isLoading!: boolean;
 
-  private _cfarOffers!: CfarOffer[];
+  private _cfarOffers!: CfarOfferCustomer[];
 
   @Input() hCSessionId!: string;
   @Input() itineraries!: CfarItinerary[];
@@ -53,11 +53,11 @@ export class CfarOfferBannerComponent extends GlobalComponent implements OnInit 
         .pipe(take(1))
         .subscribe(
           (cfarOffers) => {
-            let results: CfarOffer[] = [];
+            let results: CfarOfferCustomer[] = [];
 
             if (cfarOffers) {
               cfarOffers.forEach(cfarOffer => {
-                results.push(ApiTranslatorUtils.modelToCamelCase(cfarOffer) as CfarOffer);
+                results.push(ApiTranslatorUtils.modelToCamelCase(cfarOffer) as CfarOfferCustomer);
               });
             }
             
@@ -65,7 +65,10 @@ export class CfarOfferBannerComponent extends GlobalComponent implements OnInit 
             this.cheapestOffer = this._getCheapestOffer(this._cfarOffers);
             this.isLoading = false;
           },
-          (error) => this.isLoading = false
+          (error: any) => {
+            console.error(error);
+            this.isLoading = false;
+          }
         );
     }
   }
@@ -74,13 +77,12 @@ export class CfarOfferBannerComponent extends GlobalComponent implements OnInit 
   // Publics Methods
   // -----------------------------------------------
 
-  public onSubmit(cfarOffer: CfarOffer): void {
+  public onSubmit(cfarOffer: CfarOfferCustomer): void {
     const dialogData = { 
       currentLang: this.currentLang,
       basePath: this.basePath,
       hCSessionId: this.hCSessionId,
-      cfarOffers: this._cfarOffers,
-      extAttributes: {}
+      cfarOffers: this._cfarOffers
     };
     const dialogConfig = DialogUtils.getDialogConfig(dialogData, this.currentTheme);
     const dialogRef = this._dialog.open(CfarOfferDialogComponent, dialogConfig);
@@ -88,14 +90,14 @@ export class CfarOfferBannerComponent extends GlobalComponent implements OnInit 
     dialogRef.afterClosed()
       .pipe(take(1))
       .subscribe(
-        (result: CfarContract) => {
+        (result: CfarContractCustomer) => {
           if (result) {
             this.offerAccepted.emit(result);
           } else {
             console.log("Close dialog");
           }
         },
-        (error) => console.log(error)
+        (error) => console.error(error)
       );
   }
 
@@ -103,15 +105,14 @@ export class CfarOfferBannerComponent extends GlobalComponent implements OnInit 
   // Privates Methods
   // -----------------------------------------------
 
-  private _buildCreateCfarOfferRequest(): CreateCfarOfferRequest {
+  private _buildCreateCfarOfferRequest(): CreateCfarOfferCustomerRequest {
     return {
       itinerary: this.itineraries,
-      requestType: RequestType.Ancillary,
-      extAttributes: {}
+      requestType: RequestType.Ancillary
     };
   }
 
-  private _buildFakePostCfarOffersResponse(): CfarOffer[] {   
+  private _buildFakePostCfarOffersResponse(): CfarOfferCustomer[] {   
     return [
       {
         id: "1ecf859e-8785-625f-8eda-198d1ce0d6c4",
@@ -186,8 +187,7 @@ export class CfarOfferBannerComponent extends GlobalComponent implements OnInit 
           "Add the flexibility to cancel your flight for any reason up to 3 hours before departure",
           "Cancel and choose between a 80% refund of your flight base fare and taxes or 100% airline travel credit",
           "Get instant resolution, no forms or claims required"
-        ],
-        extAttributes: {}
+        ]
       },
       {
         id: "1ecf859e-8785-625f-8eda-198d1ce0d6c5",
@@ -262,8 +262,7 @@ export class CfarOfferBannerComponent extends GlobalComponent implements OnInit 
           "Add the flexibility to cancel your flight for any reason up to 3 hours before departure",
           "Cancel and choose between a 100% refund of your flight base fare and taxes or 100% airline travel credit",
           "Get instant resolution, no forms or claims required"
-        ],
-        extAttributes: {}
+        ]
       }
     ];
   }