+++ projects/angular-sdk/src/services/hopper-proxy.service.ts
@@ -1,7 +1,7 @@
 import { HttpClient } from "@angular/common/http";
 import { Injectable } from "@angular/core";
 import { Observable } from "rxjs";
-import { CancelForAnyReasonCFARService, CfarContract, CfarContractExercise, CfarOffer, CreateCfarContractExerciseRequest, CreateCfarContractRequest, CreateCfarOfferRequest } from "../apis/hopper-cloud-airline/v1";
+import { CfarContract, CfarContractCustomer, CfarOffer, CfarOfferCustomer, CreateCfarContractCustomerRequest, CreateCfarOfferCustomerRequest, CreateRefundAuthorizationRequest, CreateRefundRecipientRequest, CreateRefundRequest, CustomerService, RefundAuthorization, RefundRecipient } from "../apis/hopper-cloud-airline/v1";
 
 @Injectable({
   providedIn: "root"
@@ -10,39 +10,53 @@ export class HopperProxyService {
 
   constructor(
     private _httpClient: HttpClient,
-    private _cancelForAnyReasonCFARService: CancelForAnyReasonCFARService
+    private _customerService: CustomerService
   ) {}
 
   // ----------------------------------------------------------
   // CFAR
   // ----------------------------------------------------------
 
-  postCfarOffers(basePath: string, hCSessionId: string, request: CreateCfarOfferRequest): Observable<CfarOffer[]> {
+  postCfarOffers(basePath: string, hCSessionId: string, request: CreateCfarOfferCustomerRequest): Observable<CfarOfferCustomer[]> {
     // Init services
     this._overrideConfiguration(basePath);
 
-    return this._cancelForAnyReasonCFARService.postCfarOffers(request, hCSessionId);
+    return this._customerService.postCustomerCfarOffers(request, hCSessionId);
   }
 
-  postCfarContractExercises(basePath: string, hCSessionId: string, request: CreateCfarContractExerciseRequest): Observable<CfarContractExercise> {
+  postCfarContracts(basePath: string, hCSessionId: string, request: CreateCfarContractCustomerRequest): Observable<CfarContractCustomer> {
     // Init services
     this._overrideConfiguration(basePath);
 
-    return this._cancelForAnyReasonCFARService.postCfarContractExercises(request, hCSessionId);
-  }
-
-  postCfarContracts(basePath: string, hCSessionId: string, request: CreateCfarContractRequest): Observable<CfarContract> {
-    // Init services
-    this._overrideConfiguration(basePath);
-
-    return this._cancelForAnyReasonCFARService.postCfarContracts(request, hCSessionId);
+    return this._customerService.postCustomerCfarContracts(request, hCSessionId);
   }
 
   getCfarContractsId(basePath: string, hCSessionId: string, id: string): Observable<CfarContract> {
     // Init services
     this._overrideConfiguration(basePath);
 
-    return this._cancelForAnyReasonCFARService.getCfarContractsId(id, hCSessionId);
+    return this._customerService.getCustomerCfarContractsId(id, hCSessionId);
+  }
+
+  postRefundAuthorizations(basePath: string, hCSessionId: string, request: CreateRefundAuthorizationRequest): Observable<RefundAuthorization> {
+    // Init services
+    this._overrideConfiguration(basePath);
+    
+    return this._customerService.postCustomerRefundAuthorizations(request, hCSessionId);
+  }
+
+  postRefundRecipients(basePath: string, hCSessionId: string, request: CreateRefundRecipientRequest): Observable<RefundRecipient> {
+    // Init services
+    this._overrideConfiguration(basePath);
+
+    return this._customerService.postCustomerRefundRecipients(request, hCSessionId);
+  }
+
+  postRefunds(basePath: string, hCSessionId: string, request: CreateRefundRequest): Observable<any> {
+    // Init services
+    this._overrideConfiguration(basePath);
+    
+    return this._customerService.postCustomerRefunds(request, hCSessionId);
   }
 
   // ----------------------------------------------------------
@@ -50,7 +64,7 @@ export class HopperProxyService {
   // ----------------------------------------------------------
 
   private _overrideConfiguration(basePath: string): void {
-    this._cancelForAnyReasonCFARService = new CancelForAnyReasonCFARService(this._httpClient, basePath, {
+    this._customerService = new CustomerService(this._httpClient, basePath, {
       selectHeaderAccept: (accepts: ['application/json']) => 'application/json',
       selectHeaderContentType: (contentsTypes: ['application/json']) => 'application/json',
       isJsonMime: (mime: 'application/json') => true